User `Toolbar` directly instead of header widget
